[Documentation] Add a couple of package.use files
This commit adds a new directory, Documentation/package.use.unsupported/ with
files for 4.2 and 4.3.  These files are regenerated with
maintainers/regenerate-files, and get their data from sets/, so they do not
otherwise need to be maintained.

NOTE TO USERS: kdeprefix is unsupported both upstream, and by Gentoo.  There
are bugs that only apply to kdeprefix, and these may not ever be fixed.
USE AT YOUR OWN RISK.