Add {R,}DEPEND on libknotificationitem
Unfortunately, some versions of portage use a build order that puts
kde-base/libknotificationitem after packages that require it.  While it
is always placed in the deptree, sometimes it is built after packages
that need it.  To ensure that this does not happen, a DEPEND and
RDEPEND on ">=kde-base/libknotificationitem-${PV}:${SLOT}[kdeprefix=]"
has been added to each package (that I know of) that links against
libknotificationitem.so.  If and when libknotificationitem is merged
into kdelibs proper, these deps will need to be dropped.

Remember: portage does *not* always install PDEPENDs immediately after
the package that PDEPENDs on them! (In fact, in the past, packages in
PDEPEND used to always be placed as late as possible in the merge
list).