Split off kde-base/kfmclient from kde-base/konqueror
In order to not depend on konqueror in kde-base/kdebase-startkde,
kfmclient needs to be installed separately.  This entails adding a new
package, disabling the building of kfmclient in kde-base/konqueror,
adding the new package to metas and sets, adding a dependency on
kde-base/kfmclient to kde-base/kdebase-startkde, and adding appropriate
blockers on older version of kde-base/konqueror to kde-base/kfmclient.

This ended up touching many files in many directories; all changes in
this commit to 4.3.0 ebuilds will probably have to be committed to CVS
at the same time.