[eclass] Add a new add_kdebase_dep() function
The add_kdebase_dep function takes 1-2 arguments, the first being the
name of the package to depend upon, and the second being (optionally)
extra USE deps.

When given 1 argument, this function will output
	!kdeprefix? ( >=kde-base/$1-${PV}[-kdeprefix] )
	kdeprefix? ( >=kde-base/$1-${PV}:${SLOT}[kdeprefix] )

And 2 arguments outputs
	!kdeprefix? ( >=kde-base/$1-${PV}[-kdeprefix,$2] )
	kdeprefix? ( >=kde-base/$1-${PV}:${SLOT}[kdeprefix,$2] )

These are the proper deps to replace
	>=kde-base/PKG-${PV}:${SLOT}[kdeprefix=,USEDEPS]
as it does not force an all-at-once upgrade of everything.

Some of the deps (the ones from kde4-base.eclass) were already in
this form, so there was an odd combination of packages that could be
installed in separate slots with -kdeprefix, and packages that couldn't.