[eclass] Add features to add_blocker() in kde4-functions.eclass
This now allows add_blocker to add just about any blocker that you would
need for kde-base/ packages (that is, kde-base/ packages blocking other
kde-base/ packages).  In its simplest form, you can use
	add_blocker pkg-name
to add a blocker on all kde-4 versions of kde-base/pkg-name.  If further
arguments are supplied, then those indicate which versions to block.
There are 2 versions with special properties: 0 indicates that no version
is to be blocked, and 4.x.50 indicates that all versions in slots <= 4.x
are to be blocked.  A version can also be specified with a slot, like
"4.2.96:4.3", indicating that for that slot, that version is to be blocked.
If no version is given with a slot (like ":4.3"), then all versions from
that slot are blocked.  By default, versions <= the given version are
blocked.  If the given version begins with "<", however, then versions < the
given version are blocked.