[kde-base/kdelibs] Add upstream patch: KFileMetaDataProvider: Don't crash if a Strigi analyzer crashes
Some Strigi analyzers are still unstable and crash under certain circumstances.
This patch moves the reading of the meta-data into a custom process instead of
doing it in a thread, which assures that a crashing Strigi analyzer does not
result in crashing Dolphin or Konqueror when hovering some items.

(Portage version: 2.1.9.45/git/Linux x86_64, signed Manifest commit with key B6C5F7DE)