[kdemultimedia] Switch back to SVN repo for now
Switch back to SVN repo for live kdemultimedia packages as long as the
git repo is broken as told on the kde-packages list by Nicolás Alvarez:

"The kdemultimedia git repositories are *broken*, and the conversion
rules are in the process of being fixed. Please release kdemultimedia
from the SVN KDE/4.8 branch. The git repositories were locked after
the brokenness was noticed, so there were even some kmix fixes done on
the SVN branch and not on the git repos."

We will just revert this commit as soon as the git repo has been fixed
by the KDE people to use git again.