[eclass] Always add aqua to IUSE
We always need the aqua useflag because otherwise we cannot = refer to it inside
add_kdebase_dep. This was always kind of a bug, but came to light with EAPI=5
(where referring to a use flag not in IUSE masks the ebuild).
The only alternative would be to prohibit using add_kdebase_dep if KDE_REQUIRED=never