[x11-themes/bespin] Improve dependencies.
It is necessary to specify Qt/X11 dependencies explicitly, since the
KDE eclass will only pull them in conditionally with USE="kde".

It's not necessary to pull in kdelibs with USE="plasma", since
this requires USE="kde" which unconditionally pulls in kdelibs.

dev-qt/qtgui[dbus] is not required since in this case it only served
to pull in qtdbus, which we now explicitly require.

Explicitly set RDEPEND="${DEPEND}" since this is required in EAPI 5.

Package-Manager: portage-2.2.7