[kde-base/plasma-workspace] Update dependencies.
drkonqi is the sole consumer of kdewebkit, and requires kxmlrpcclient which is
currently unreleased. Previously, building drkonqi was optional because of
this. Now that that a bundled version is included, drkonqi is built
unconditionally causing kdewebkit to now really be required.

This needs to be reviewed in the future when kxmlrpcclient has a release.

This closes bug #511494, but does not solve the original underlying issue (which
no longer exists).

Upstream commit: 29684cba61d65e6fbca4b0dd23dc1bbd9e50fc00
Upstream commit: 80674e133fcd4a0c1b335cff6824c1786b2d8a48

Package-Manager: portage-2.2.10