kde-misc/rsibreak: Add a live ebuild for the KF5 version of rsibreak
I've changed the ${LICENSE} to show that it's actually GPLv2 or later
because that's what the source files say, based on a quick grep. There's
no support for translations so far because I have no idea how they work
in "kde5".

The blocker is in place because the eclass apparently only works for
kde-apps/, not kde-misc/.