RFC dev-libs/libunity: added 7.1.4
when libunity is installed the taskmanager entries of chromium (and most
likely other applications) get additional information. For chromium that is
download progress and number of active downloads

This ebuild is trivially copied from the unity overlay and needs additional
work. I just wanted to ask you how we should deal with the ubuntu versioning
scheme.

I will tidy up the ebuild later, but you can play with it already