cmake.eclass: do not append -DNDEBUG to CPPFLAGS
The NDEBUG macro turns the assert() function into a noop. This gives a
small performance boost, but may allow subtle programming errors to go
unnoticed.

This code was added back in 2008, when we started passing
-DCMAKE_BUILD_TYPE=None instead of Release or Debug. It probably tries
to mimic a default behavior of Release type builds.

Other common build systems do not do this by default. For example,
autoconf's AC_HEADER_ASSERT macro only sets NDEBUG if --disable-assert
is passed to configure (it defaults to enabled).

It is better to let users add this to CPPFLAGS themselves if they really
want to save those few CPU cycles.

Signed-off-by: Mike Gilbert <floppym@gentoo.org>
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>