kde-apps/libkgapi: skip buggy tests
We want to report these failures upstream but I didn't
think ahead and upgraded to newer Qt 5.15.3 (not yet in wider
distribution yet in Gentoo) and some of the newer failures I hit
may be attributed to that.

So, let's skip them for now, and circle back. Other distros
(Exherbo, Alpine) are hitting these failures anyway so we know it's not
Gentoo-specific.

Bug: https://bugs.gentoo.org/832709
Signed-off-by: Sam James <sam@gentoo.org>