ecm.eclass: Restrict ECM_TEST="optional" and punting dep to _KFSLOT=5
In dev-qt/qtbase:6, Qt6Test is always provided so it makes no sense to
go through the trouble of ripping out the build dependency.

We should still aim to avoid needless building of tests if not requested,
either by detecting unconditional use of ecm_add_test or expose existing
force* logic to ebuild developers to catch those as errors.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>